backend/requirements/base.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Real-time & WebSockets
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1

# Database
psycopg2-binary==2.9.9
aiosqlite==0.19.0

# Authentication
dj-rest-auth==5.0.2
django-allauth==0.60.0
djangorestframework-simplejwt==5.3.0

# WebRTC
aiortc==1.6.0
av==11.0.0

# Async & Background Tasks
celery==5.3.4
django-celery-beat==2.5.0

# Media & Storage
pillow==10.1.0
django-storages==1.14.2
boto3==1.34.0

# Utilities
python-dotenv==1.0.0
django-environ==0.11.2
channels==4.0.0

# Email
sendgrid==6.10.0

# Security
django-ratelimit==4.1.0
bcrypt==4.1.2
cryptography==41.0.7